Move "open project" to the recent projects
